requirements.txt
annotated-types==0.7.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
Pillow==10.3.0
pydantic==2.8.2
pydantic_core==2.20.1
SQLAlchemy==2.0.31
typing_extensions==4.12.2
Werkzeug==3.0.3
